Tasks are sorted according to order of adding
People tend to want to know their deadlines first to check it off before the deadline. Compared to tasks, which can be completed as and when they wish to. Events are simply things to attend
Add sort command which will sort the tasklist according to deadlines, followed by tasks then events
This allows users to quickly scan their list of deadlines

Future improvements to the function will include sorting according to closest date and time to the present